Update existing tests with company cut, reminder, and memebrs as null
